backend/requirements.txt
bcrypt==4.1.3
alembic==1.17.2
psycopg2-binary==2.9.10
anyio==4.11.0
asyncpg==0.31.0
httpx==0.28.1
openai>=1.40,<3
fastapi==0.122.0
ormar==0.23.1
passlib[bcrypt]==1.7.4
prometheus-fastapi-instrumentator==7.1.0
pydantic==2.12.4
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
SQLAlchemy==2.0.44
starlette==0.50.0
uvicorn==0.38.0
pytest==8.4.2
respx==0.22.0
